requirements.txt
torch
torchvision
numpy
opencv-python
matplotlib
scikit-learn
tqdm
Pillow